1.2.0 Update
This update finally fixes the issue where you could not reload the config on Spigot or any forks.
The reload command is now /bhook reload with the permissions remaining as bhook.reload
Also added punishment IDs to embeds. This is the same ID as used by litebans, so would help servers that use punishment IDs as a method of management to more easily find the associated punishment.